+++ e2e/cypress.config.js
@@ -3,21 +3,91 @@ const { exec, spawn } = require("child_process");
 const cypressFailFast = require("cypress-fail-fast/plugin.js");
 let firebackProcess; // Store the Fireback process reference
 
-let CWD = "";
-let BINARY = "";
+let BINARY = "/Users/ali/work/fireback/app";
+let CWD = "/Users/ali/work/fireback";
+const PORT = 7793;
+let DB_VENDOR = "sqlite";
+const isGitHubActions = !!process.env.GITHUB_ACTIONS;
+
+if (isGitHubActions) {
+  // BINARY = "/usr/local/bin/fireback";
+  BINARY = "/usr/local/bin/fireback";
+  CWD = "/home/runner/work/fireback";
+  // CWD = "/home/runner/work/fireback-private/fireback-private";
+}
+
+console.log(BINARY);
+console.log(CWD);
+
+const execAsync = (cmd, CWD = "") => {
+  return new Promise((resolve, reject) => {
+    exec(cmd, { cwd: CWD }, (error, stdout, stderr) => {
+      if (error) {
+        return reject(error + " --- " + cmd + "----" + CWD);
+      }
+      resolve(stdout || stderr);
+    });
+  });
+};
 
 module.exports = defineConfig({
   video: true,
   chromeWebSecurity: false,
   env: {
     GITHUB_ACTIONS: process.env.GITHUB_ACTIONS,
+    PORT: PORT,
   },
   e2e: {
     setupNodeEvents(on, config) {
       cypressFailFast(on, config);
+
+      on("task", {
+        log(message) {
+          console.log(message);
+          return null;
+        },
+      });
+
+      on("task", {
+        view(url) {
+          return cy.visit(url);
+        },
+        async dbcon(cmd) {
+          const vendor = process.env.DB_TYPE;
+          console.log(100000, vendor);
+          const dbname = `test_agent_${new Date().getTime()}`;
+
+          if (vendor === "mysql") {
+            try {
+              await execAsync(`${BINARY} config db-vendor set mysql`, CWD);
+              await execAsync(
+                `${BINARY} config db-dsn set "root:root@tcp(localhost:3306)/fireback_test?charset=utf8mb4&parseTime=True&loc=Local"`,
+                CWD
+              );
+              await execAsync(`${BINARY} migration apply`, CWD);
+
+              return true;
+            } catch (err) {
+              console.error("setup mysql failed:", err);
+            }
+          } else {
+            await execAsync(`${BINARY} config db-vendor set sqlite`, CWD);
+            await execAsync(
+              `${BINARY} config db-name set /tmp/${dbname}.db`,
+              CWD
+            );
+            await execAsync(`${BINARY} migration apply`, CWD);
+
+            return true;
+          }
+
+          return false;
+        },
+      });
       on("task", {
         exec(cmd) {
           return new Promise((resolve, reject) => {
+            cmd = BINARY + " " + cmd;
             console.log("Running:", cmd, " on: ", CWD);
             exec(cmd, { cwd: CWD }, (error, stdout, stderr) => {
               if (error) {
@@ -31,6 +101,7 @@ module.exports = defineConfig({
       on("task", {
         execSupress(cmd) {
           return new Promise((resolve, reject) => {
+            cmd = BINARY + " " + cmd;
             console.log("Running:", cmd, " on: ", CWD);
             exec(cmd, { cwd: CWD }, (error, stdout, stderr) => {
               if (error) {
@@ -46,7 +117,7 @@ module.exports = defineConfig({
         startFireback() {
           return new Promise((resolve, reject) => {
             console.log("Starting Fireback...");
-            firebackProcess = spawn(`PORT=4502 ${BINARY}`, ["start"], {
+            firebackProcess = spawn(`PORT=${PORT} ${BINARY}`, ["start"], {
               stdio: "inherit",
               shell: true,
               cwd: CWD,
@@ -89,3 +160,8 @@ module.exports = defineConfig({
     },
   },
 });
+
+// You can change the binary and cwd using this directly in the tests
+// beforeEach(() => {
+//   cy.task("execCwd", { cwd, binary });
+// });