+++ modules/workspaces/queries/WorkspaceCommonPivot.sqlite.dynoCounter.sql
@@ -1,4 +1,4 @@
 select
     count(*) total_items
 from
-    fb_workspace_entities
+    workspace_entities